requirements.txt
pyserial>=3.5
fprime-gds>=3.3.0 